requirements.txt
# Core
coco-backends==0.0.1
coco-common==0.0.1
coco-contract==0.0.1
django==1.8.3
django-admin-conf-vars==0.3.1
djangorestframework==3.1.1
# apt-get -y install python-psycopg2
# psycopg2

# Admin
djangocms-admin-style==0.2.7
docutils==0.12

# Web
django-widget-tweaks==1.3
